requirements.txt
# Dependências do Projeto Atrio

# Core
python-dotenv==1.0.0

# PostgreSQL (conexão direta ao Supabase)
psycopg2-binary==2.9.9

# Git automation já usa subprocess (built-in)
